91.2.Handling Requests with Async Await(Section 8: Making API Requests with React)
So let's now take a look at the alternate method.

Like I said this alternate method for getting a little notification of when we get this response back

is a little bit easier to write out.

All right so I'm going to remove the Dathan statement.

I'm going to find my own search submit function and I'm going to market with the async keyword like

so by putting the async keyword in front of the function name.

That allows us to use the async or wait syntax inside this function.

What that really means is that in front of the network request that we make to ex-U.S. right there I'm

going to put a wait keyword and then in front of that I'm going to assign a new variable called response

like so.

So in total I have Konst response equals a wait ASIO's and then down here after that entire request

I can count the log out response that data got results like so.

All right before I tell you anything about this syntax right here let's save the file and then run it

again very quickly so I'll put in a search term of cars and hit enter and I see again I get my array

of results right here just as we did previously.

So we have the promise space syntax that uses a then statement that will be called Anytime the request

gets completed.

Alternatively we can use this async await syntax.

In my opinion a sink a weight is far easier to understand and to write out in code than making use of

the promise.

When do you want to use the async syntax.

You just put that async keyword in front of the method name you find whatever is returning or whatever

is taking some time to resolve.

In this case it is the network request with X-post get.

And we put the a week key word in front of that and then whatever gets returned from this entire statement

right here.

We will assign to some variable as we usually do and then we can freely work with that variable later

on.

We're going to be using this async 08 syntax many times throughout this course.

So we'll get a lot of practice with it and some practice with some variations that add in some pretty

good error handling with it as well.

All right.

So we now are able to get our list of images here.

Let's take a quick pause.

When we come back the next section we're going to update our app component to have a state property

and make sure that we set our lists of images on that state.

So a quick pause and I'll see you in just a minute.