99.2.Implementing Keys in Lists<(Section 9: Building Lists of Records)
Now one last thing I want to do here.

You'll notice that we are referencing the image variable three times inside this mapping function.

So rather than keeping on referencing that variable again and again and again we can just do restructure

out the three properties we care about inside of this inner function description ID and you or else

to do so I'll find that image argument.

I'm going to delete it.

I'll put down a set of parentheses and a set of curly braces inside that.

And then I can structure off description ID.

And you are L's like so and then I can remove the image and the Daut the ID and the DOT and the image

and the dot or see me.

I didn't mean to say id and dot that was image and Dot.

OK.

That's it.

Little bit shorter than what we had before.

So going to save this will do one more quick test.

Make sure we don't have any warnings or anything like that.

So search for cars and no warnings no error messages.

OK.

So that is pretty much it for showing a list of elements we've now got our list of car pictures or whatever

it is that you want to search for.

I don't know why.

As example I kept on saying like trees and plants and flowers.

I don't know why that came to mind but you know what that will do flowers anyways.

So we definitely got our list of images here which is fantastic.

But that's really just the one of our application.

Remember we had said we just want to get the list of images to show up so we're not going to say cool

looks good.

Now we're going to start to do some extra credit.

All right.

So what we're going go through starting in the next video is stuff that you don't necessarily have to

know to understand react and how it works.

We're going to go over a couple of topics that are decidedly penned not related to react but nonetheless

I think it's going to be some pretty interesting stuff.

We're going to figure out how to style these images far better than they are currently displayed.

So quick pause here and we'll continue in the next video.