100.2. Grid CSS(Section 10: Using Ref's for DOM Access)
So in order to make sure that there's a little bit of standoff between those images we can add in an

additional rule of grid ash gap 10 pixels like so.

So that's going to make sure that each of these different cells have a 10 pixel load spacing between

them on both the sides and in between top and bottom.

Let me save that CSSA file.

And if I flip back over I'll see that I now get a little bit of separation in there.

Now when you do that inspect Now you can really start to see those cells come together.

All right.

So next on Tell you a little more about the grid template columns rule.

Essentially what this does is it creates a set number of columns autofill right here means that the

CSSA roll or the grid CSSA system is going to automatically decide how many columns to insert on each

year.

So in this case it decided to put in two and then the second argument right here minimax 250 pixels

one FFR means that each one of those different columns is going to be at minimum 250 pixels wide and

then at most and this one's a little bit challenging to understand one as far as space.

Basically a maximum allocation of space one for in this context means that for every column that we

creates We want them to all be equally sized.

So in this case it's kind of challenging to communicate exactly what one does.

Essentially if we had multiple columns being created and we were manually creating each one we could

assign a different value to them and then some different color or some of those columns might have a

greater or smaller width assigned to them automatically.

All right now adding in that column thing definitely looks like it improved things.

You know we got a little bit of standoff here but remember the original idea was to make sure that we've

got this like really nice kind of closely fitted set of images.

So let's take a quick pause right here.

When we come back the next section we're going to do a little bit more work on this grid system and

see if we can't get these things to fit together a little bit more nicely than they currently are.